test(Car): rewrite test with StaticRandomGenerator
As Car.moveOrStop() use random value, it is hard to set the proper
result state of Car after calling the method.

To solve this problem, the constructor of Car now accept
RandomGenerator. The test function can control what value would return
for the moveOrStop() by passing StaticRandomGenerator to the constructor
of Car and set staticNumber of StaticRandomGenerator to the number
expected to be returned.