refactor(Car): remove toString() override
Car had toString() override to return pretty print string of it's name
and progress (e.g. "bentley : --"), but it seems not proper to implement
presentation functionality to Car. Car is only responsible to store the
status of car and execute alloweded tasks.

Implementation of pretty print car's status is now move to the
Application.